separate cats reduction features into its own file
